Remove spinlock, change to i2c_readwrite where appropriate , and consolidate i2c read and write calls.

+++ include/nuttx/sensors/apds9922.h
@@ -191,10 +191,10 @@ struct apds9922_als_setup_s
 {
   int      rate;        /* als measurement rate              */
   int      res;         /* als resolution                    */
-  int      gain;        /* als gain                          */
+  uint8_t  gain;        /* als gain                          */
   struct   adps9922_als_thresh
            thresh;      /* Upper and lower thresholds        */
-  int      thresh_var;  /* threshold variation               */
+  uint8_t  thresh_var;  /* threshold variation               */
   int      int_mode;    /* Interrupt mode                    */
   uint8_t  persistance; /* Num events before interrupt       */
   uint32_t als_factor;  /* Lux correction factor applied     */
@@ -227,7 +227,7 @@ struct apds9922_ps_setup_s
 
 struct apds9922_ps_data
 {
-  uint16_t ps;          /* Current prximity measure          */
+  uint16_t ps;          /* Current proximity measure         */
   bool     close;       /* Object is far (false) or close    */
 };
